+++ scene/2d/collision_polygon_2d.h
@@ -75,7 +75,7 @@ public:
 
 	Rect2 _edit_get_rect() const override;
 	bool _edit_use_rect() const override;
-	bool _edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const override;
+    bool _edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const override;
 
     StringName get_configuration_warning() const override;
