+++ scene/2d/collision_shape_2d.h
@@ -55,7 +55,7 @@ protected:
 	static void _bind_methods();
 
 public:
-	bool _edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const override;
+    bool _edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const override;
 
 	void set_shape(const Ref<Shape2D> &p_shape);
 	Ref<Shape2D> get_shape() const;
