+++ scene/2d/light_occluder_2d.h
@@ -58,7 +58,7 @@ protected:
 
 public:
 	virtual Rect2 _edit_get_rect() const;
-	virtual bool _edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const;
+    virtual bool _edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const;
 
 	void set_polygon(const PoolVector<Vector2> &p_polygon);
 	PoolVector<Vector2> get_polygon() const;
@@ -91,7 +91,7 @@ protected:
 
 public:
 	Rect2 _edit_get_rect() const override;
-	bool _edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const override;
+    bool _edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const override;
 
 	void set_occluder_polygon(const Ref<OccluderPolygon2D> &p_polygon);
 	Ref<OccluderPolygon2D> get_occluder_polygon() const;
