+++ scene/2d/line_2d.cpp
@@ -73,7 +73,7 @@ bool Line2D::_edit_use_rect() const {
     return true;
 }
 
-bool Line2D::_edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const {
+bool Line2D::_edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const {
 
     const real_t d = _width / 2 + p_tolerance;
     PoolVector<Vector2>::Read points = _points.read();
