+++ scene/2d/path_2d.h
@@ -49,7 +49,7 @@ protected:
 public:
 	Rect2 _edit_get_rect() const override;
 	bool _edit_use_rect() const override;
-	bool _edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const override;
+    bool _edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const override;
 
 	void set_curve(const Ref<Curve2D> &p_curve);
 	Ref<Curve2D> get_curve() const;
