+++ scene/3d/area.cpp
@@ -130,7 +130,7 @@ void Area::_body_enter_tree(ObjectID p_id) {
 
     E->second.in_tree = true;
     emit_signal(SceneStringNames::get_singleton()->body_entered,  Variant(node));
-    for (int i = 0; i < E->second.shapes.size(); i++) {
+    for (size_t i = 0; i < E->second.shapes.size(); i++) {
 
         emit_signal(SceneStringNames::get_singleton()->body_shape_entered, p_id, Variant(node), E->second.shapes[i].body_shape, E->second.shapes[i].area_shape);
     }
@@ -146,7 +146,7 @@ void Area::_body_exit_tree(ObjectID p_id) {
     ERR_FAIL_COND(!E->second.in_tree)
     E->second.in_tree = false;
     emit_signal(SceneStringNames::get_singleton()->body_exited, Variant(node));
-    for (int i = 0; i < E->second.shapes.size(); i++) {
+    for (size_t i = 0; i < E->second.shapes.size(); i++) {
 
         emit_signal(SceneStringNames::get_singleton()->body_shape_exited, p_id, Variant(node), E->second.shapes[i].body_shape, E->second.shapes[i].area_shape);
     }
@@ -242,7 +242,7 @@ void Area::_clear_monitoring() {
             if (!E.second.in_tree)
                 continue;
 
-            for (int i = 0; i < E.second.shapes.size(); i++) {
+            for (size_t i = 0; i < E.second.shapes.size(); i++) {
 
                 emit_signal(SceneStringNames::get_singleton()->body_shape_exited, E.first, Variant(node), E.second.shapes[i].body_shape, E.second.shapes[i].area_shape);
             }
@@ -272,7 +272,7 @@ void Area::_clear_monitoring() {
             if (!E.second.in_tree)
                 continue;
 
-            for (int i = 0; i < E.second.shapes.size(); i++) {
+            for (size_t i = 0; i < E.second.shapes.size(); i++) {
 
                 emit_signal(SceneStringNames::get_singleton()->area_shape_exited, E.first, Variant(node), E.second.shapes[i].area_shape, E.second.shapes[i].self_shape);
             }
@@ -323,7 +323,7 @@ void Area::_area_enter_tree(ObjectID p_id) {
 
     E->second.in_tree = true;
     emit_signal(SceneStringNames::get_singleton()->area_entered, Variant(node));
-    for (int i = 0; i < E->second.shapes.size(); i++) {
+    for (size_t i = 0; i < E->second.shapes.size(); i++) {
 
         emit_signal(SceneStringNames::get_singleton()->area_shape_entered, p_id, Variant(node), E->second.shapes[i].area_shape, E->second.shapes[i].self_shape);
     }
@@ -339,7 +339,7 @@ void Area::_area_exit_tree(ObjectID p_id) {
     ERR_FAIL_COND(!E->second.in_tree)
     E->second.in_tree = false;
     emit_signal(SceneStringNames::get_singleton()->area_exited, Variant(node));
-    for (int i = 0; i < E->second.shapes.size(); i++) {
+    for (size_t i = 0; i < E->second.shapes.size(); i++) {
 
         emit_signal(SceneStringNames::get_singleton()->area_shape_exited, p_id, Variant(node), E->second.shapes[i].area_shape, E->second.shapes[i].self_shape);
     }
