+++ scene/resources/convex_polygon_shape_2d.h
@@ -43,7 +43,7 @@ protected:
 	static void _bind_methods();
 
 public:
-	bool _edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const override;
+    bool _edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const override;
 
 	void set_point_cloud(const Vector<Vector2> &p_points);
 	void set_points(const Vector<Vector2> &p_points);
